added bluetooth connection & DetailView

+++ IOSBle-RedBear/BLEFramework/BLE/BLE.m
@@ -184,7 +184,7 @@ static int rssi = 0;
     if (self.CM.state != CBCentralManagerStatePoweredOn)
     {
         NSLog(@"CoreBluetooth not correctly initialized !");
-        NSLog(@"State = %d (%s)\r\n", self.CM.state, [self centralManagerStateToString:self.CM.state]);
+        NSLog(@"State = %ld (%s)\r\n", (long)self.CM.state, [self centralManagerStateToString:self.CM.state]);
         return -1;
     }
     
@@ -578,5 +578,6 @@ static bool done = false;
         [[self delegate] bleDidUpdateRSSI:activePeripheral.RSSI];
     }
 }
+
 
 @end
